requirements/_base.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
django-sekizai
psycopg2-binary
asgiref==3.3.1
cairocffi==1.1.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.12.3
cssselect2==0.2.2
defusedxml==0.6.0
Django==3.0.13
django-appconf==1.0.3
django-crispy-forms==1.11.1
django-elasticsearch-dsl==7.1.4
django-haystack==3.0
django-imagekit==4.0.2
django-json-ld==0.0.4
django-qr-code==2.1.0
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
html5lib==1.0.1
pilkit==2.0
Pillow==8.1.1
pycparser==2.19
Pyphen==0.9.5
python-dateutil==2.8.0
Open-Weather-API
ip2geotools
django-ipware